requirements.txt
Django==1.5.12
Pillow==2.4.0
Pygments==1.6
South==0.8.2
Werkzeug==0.9.4
argparse==1.2.1
django-annoying==0.7.7
django-extensions==1.1.1
django-fack==1.1
django-file-resubmit==0.4.1
django-markupfield==1.2.0
docutils==0.11
humanize==0.5
pygeoip==0.2.7
six==1.5.2
sqlparse==0.1.11
wsgiref==0.1.2
stripe
constantcontact==0.3.1
requests
